# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lwang Black API",
    "description": "E-commerce backend API for Lwang Black specialty coffee. Supports multi-currency payments (Stripe, PayPal, Khalti, eSewa, Nabil Bank), global logistics (Shippo, DHL, FedEx), admin dashboard, and notification services.",
    "version": "1.0.0",
    "contact": {
      "email": "brewed@lwangblack.co",
      "url": "https://www.lwangblack.co"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" },
    { "url": "https://api.lwangblack.co", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "category": { "type": "string", "enum": ["coffee", "accessories", "bundles", "apparel"] },
          "description": { "type": "string" },
          "prices": { "type": "object" },
          "stock": { "type": "integer" },
          "status": { "type": "string", "enum": ["active", "draft", "archived"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List all products",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Product list with Redis caching" }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a product",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": { "201": { "description": "Product created" } }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID or slug",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update a product",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Archive a product (owner only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Archived" } }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get store settings",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Key-value settings" } }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update store settings (owner only)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Settings saved" } }
      }
    },
...
  }
}